Encapsulate preview card component
